add movie series modul

+++ src/main/java/net/kleditzsch/SmartHome/view/movie/user/moviebox/MovieBoxMovieFormServlet.java
@@ -25,9 +25,14 @@ import javax.servlet.http.Part;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
+import java.net.URI;
+import java.net.http.HttpClient;
+import java.net.http.HttpRequest;
+import java.net.http.HttpResponse;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.time.Duration;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.Arrays;
@@ -154,7 +159,7 @@ public class MovieBoxMovieFormServlet extends HttpServlet {
         //Optionale Parameter
         ID movieId = null;
         Part cover = null;
-        String coverPath = null;
+        String coverPath = null, coverUrl = null;
         List<ID> actors = null, directors = null;
 
         FormValidation form = FormValidation.create(req);
@@ -187,6 +192,10 @@ public class MovieBoxMovieFormServlet extends HttpServlet {
         if(form.fieldNotEmpty("coverPath")) {
 
             coverPath = form.getString("coverPath", "Cover Pfad", 3, 100);
+        }
+        if(form.fieldNotEmpty("coverUrl")) {
+
+            coverUrl = form.getString("coverUrl", "Cover URL", 3, 1000);
         }
 
         //IDs prüfen
@@ -294,6 +303,67 @@ public class MovieBoxMovieFormServlet extends HttpServlet {
                         }
 
                         movie.setCoverFile(filename);
+                    } else if (coverUrl != null) {
+
+                        //Cover aus dem Internet herunterladen
+                        Path tmpDir = Paths.get("upload/tmp");
+                        if(!Files.exists(tmpDir)) {
+
+                            Files.createDirectories(tmpDir);
+                        }
+
+                        try {
+
+                            HttpClient client = HttpClient.newBuilder()
+                                    .version(HttpClient.Version.HTTP_2)
+                                    .followRedirects(HttpClient.Redirect.NORMAL)
+                                    .build();;
+
+                            HttpRequest request = HttpRequest.newBuilder()
+                                    .uri(URI.create(coverUrl))
+                                    .timeout(Duration.ofSeconds(5))
+                                    .GET()
+                                    .build();
+
+                            Path tmpFile = tmpDir.resolve("tmpCover");
+                            HttpResponse.BodyHandler<Path> asFile = HttpResponse.BodyHandlers.ofFile(tmpFile);
+                            HttpResponse<Path> response = client.send(request, asFile);
+
+                            if(response.statusCode() == 200) {
+
+                                //neuen Dateinamen erstellen und COntent Type prüfen
+                                String filename = ID.create().get();
+                                switch(response.headers().allValues("Content-Type").get(0)) {
+
+                                    case "image/jpeg":
+
+                                        filename += ".jpeg";
+                                        break;
+                                    case "image/png":
+
+                                        filename += ".png";
+                                        break;
+                                    case "image/gif":
+
+                                        filename += ".gif";
+                                        break;
+
+                                    default:
+
+                                        //Ungültiger Dateityp
+                                        Files.delete(tmpFile);
+                                        throw new IllegalStateException();
+                                }
+
+                                //Datei in Cover Ordner verschieben und Dateinem im Filmobjekt speichern
+                                Files.move(tmpFile, uploadDir.resolve(filename));
+                                movie.setCoverFile(filename);
+                            } else {
+
+                                //Download fehleschlagen
+                                throw new IllegalStateException("Download fehlgeschlagen");
+                            }
+                        } catch (InterruptedException e) {}
                     } else if (coverPath != null && !tmdbApiKey.isEmpty()) {
 
                         //Cover von The Movie DB herunterladen
@@ -423,11 +493,84 @@ public class MovieBoxMovieFormServlet extends HttpServlet {
                             //altes Logo löschen
                             if(movie.getCoverFile() != null && movie.getCoverFile().length() > 0) {
 
-                                Files.delete(uploadDir.resolve(movie.getCoverFile()));
+                                try {
+
+                                    Files.delete(uploadDir.resolve(movie.getCoverFile()));
+                                } catch (Exception e) {}
                             }
 
                             //Dateiname des neuen Logos setzen
                             movie.setCoverFile(filename);
+                        } else if (coverUrl != null) {
+
+                            //Cover aus dem Internet herunterladen
+                            Path tmpDir = Paths.get("upload/tmp");
+                            if(!Files.exists(tmpDir)) {
+
+                                Files.createDirectories(tmpDir);
+                            }
+
+                            try {
+
+                                HttpClient client = HttpClient.newBuilder()
+                                        .version(HttpClient.Version.HTTP_2)
+                                        .followRedirects(HttpClient.Redirect.NORMAL)
+                                        .build();;
+
+                                HttpRequest request = HttpRequest.newBuilder()
+                                        .uri(URI.create(coverUrl))
+                                        .timeout(Duration.ofSeconds(5))
+                                        .GET()
+                                        .build();
+
+                                Path tmpFile = tmpDir.resolve("tmpCover");
+                                HttpResponse.BodyHandler<Path> asFile = HttpResponse.BodyHandlers.ofFile(tmpFile);
+                                HttpResponse<Path> response = client.send(request, asFile);
+
+                                if(response.statusCode() == 200) {
+
+                                    //neuen Dateinamen erstellen und COntent Type prüfen
+                                    String filename = ID.create().get();
+                                    switch(response.headers().allValues("Content-Type").get(0)) {
+
+                                        case "image/jpeg":
+
+                                            filename += ".jpeg";
+                                            break;
+                                        case "image/png":
+
+                                            filename += ".png";
+                                            break;
+                                        case "image/gif":
+
+                                            filename += ".gif";
+                                            break;
+
+                                        default:
+
+                                            //Ungültiger Dateityp
+                                            Files.delete(tmpFile);
+                                            throw new IllegalStateException();
+                                    }
+
+                                    //altes Logo löschen
+                                    if(movie.getCoverFile() != null && movie.getCoverFile().length() > 0) {
+
+                                        try {
+
+                                            Files.delete(uploadDir.resolve(movie.getCoverFile()));
+                                        } catch (Exception e) {}
+                                    }
+
+                                    //Datei in Cover Ordner verschieben und Dateinem im Filmobjekt speichern
+                                    Files.move(tmpFile, uploadDir.resolve(filename));
+                                    movie.setCoverFile(filename);
+                                } else {
+
+                                    //Download fehleschlagen
+                                    throw new IllegalStateException("Download fehlgeschlagen");
+                                }
+                            } catch (InterruptedException e) {}
                         } else if (coverPath != null && !tmdbApiKey.isEmpty()) {
 
                             //Cover von The Movie DB herunterladen
